log scale base as double >= 1.1
git-svn-id: svn://svn.code.sf.net/p/qwt/code/trunk/qwt@1318 29e17f8d-cf37-471e-a288-72cc03a76eda

+++ src/qwt_scale_engine.h
@@ -186,7 +186,7 @@ protected:
 class QWT_EXPORT QwtLogScaleEngine: public QwtScaleEngine
 {
 public:
-    QwtLogScaleEngine( int base = 10 );
+    QwtLogScaleEngine( double base = 10.0 );
     virtual ~QwtLogScaleEngine();
 
     virtual void autoScale( int maxSteps,
